refactor: move VI testing into its own file
